Fix bugprone-narrowing-conversions warnings in core/vnl
Add explicit static_cast<> conversions throughout core/vnl to resolve
all clang-tidy bugprone-narrowing-conversions warnings. Covers 35 files
including source, headers, and .hxx template implementations.

Fixes test failure: vul_vul_test_driver

Key patterns fixed:
- size_t/.size() → int: static_cast<int>()
- unsigned int → int at function call sites
- long → int in getter return values
- size_t → double in arithmetic expressions
- long double → double for complex number operations
- unsigned short arithmetic → short for vnl_bignum::resize() calls
- long (rational int_type) → double: replaced 0.5*den comparisons
  with integer arithmetic (2*(num%den) >= den)
- vnl_matrix.hxx vnl_inplace_transpose: added local int aliases
  im_/in_ for signed arithmetic on unsigned m/n parameters
- Sparse matrix inner loop vars changed from unsigned int to int
  (sparse_vector holds pair<int,int>)

+++ core/vnl/algo/vnl_gaussian_kernel_1d.h
@@ -29,7 +29,7 @@ public:
   int
   width() const
   {
-    return vec_.size();
+    return static_cast<int>(vec_.size());
   }
   double
   operator[](int i) const
